% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_plot.R
\name{annotation_plot}
\alias{annotation_plot}
\title{A function to draw a density plot comparing cases and controls by PRS}
\usage{
annotation_plot(prsModel)
}
\arguments{
\item{prsModel}{A \code{prsModel} object, created by \code{\link[=prs_models]{prs_models()}}.}
}
\value{
A ggplot object.
}
\description{
A function to draw a density plot comparing cases and controls by PRS
}
\examples{
\dontrun{
prs_models(PRSdata, exposure="prs",outcome="disease", covariates=c("age" ,"sex"), nquantiles=10) |>
  annotation_plot()
}
}
