% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_sup.R
\name{exp_sup}
\alias{exp_sup}
\title{Format a P value in markdown or HTML}
\usage{
exp_sup(x, pvalue = 2, boldsig = FALSE, md = FALSE, sig = 0.05)
}
\arguments{
\item{x}{A numeric vector with one element}

\item{pvalue}{Number of decimal places needed}

\item{boldsig}{Whether to make bold significant P values}

\item{md}{Whether to use markdown or HTML}

\item{sig}{A number vector of the significance level}
}
\description{
Format a P value in markdown or HTML
}
\examples{
exp_sup(0.00000456)
exp_sup(0.002, md=TRUE, boldsig=TRUE)
}
