% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggroc_plot.R
\name{ggroc_plot}
\alias{ggroc_plot}
\title{A function to draw a density plot comparing cases and controls by PRS}
\usage{
ggroc_plot(PRSdata, exposure, outcome, covariates, comparison = NA, nquantiles)
}
\arguments{
\item{PRSdata}{A dataframe which includes an exposure and outcome}

\item{exposure}{A character string containing the name of the exposure variable}

\item{outcome}{A character string containing the name of the outcome variable}

\item{covariates}{Character vector containing names of covariates}

\item{comparison}{Character variable containing name a secondary PRS}

\item{nquantiles}{Number of quantiles}
}
\value{
A ggplot object.
}
\description{
A function to draw a density plot comparing cases and controls by PRS
}
\examples{
\dontrun{
ggroc_plot(PRSdata, exposure="prs",outcome="disease",covariates=c("age", "sex"),nquantiles=10)
}
}
