% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_density_plot.R
\name{prs_density_plot}
\alias{prs_density_plot}
\title{A function to draw a density plot comparing cases and controls by PRS}
\usage{
prs_density_plot(PRSdata, exposure, outcome)
}
\arguments{
\item{PRSdata}{A dataframe which includes an exposure and outcome}

\item{exposure}{A character string containing the name of the exposure variable}

\item{outcome}{A character string containing the name of the outcome variable}
}
\value{
A ggplot object.
}
\description{
A function to draw a density plot comparing cases and controls by PRS
}
\examples{
\dontrun{
prs_density_plot(PRSdata, exposure="prs",outcome="disease")
}
}
