% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_models_NEW.R
\name{prs_models_NEW}
\alias{prs_models_NEW}
\title{Polygenic risk score modelling functions}
\usage{
prs_models_NEW(
  PRSdata,
  exposure,
  outcome,
  covariates,
  comparison = NA,
  nquantiles
)
}
\arguments{
\item{PRSdata}{Data frame containing exposure and outcome}

\item{exposure}{Character variable containing name of polygenic risk score variable}

\item{outcome}{Character variable containing name of binary outcome variable}

\item{covariates}{Character variable containing names of covariates}

\item{comparison}{Character variable containing name a secondary PRS}

\item{nquantiles}{Number of quantiles}
}
\description{
Polygenic risk score modelling functions
}
\examples{
prs_models_NEW(PRSdata,exposure="prs",outcome="disease",covariates="age,sex",nquantiles=10)
}
