% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSdata.R
\docType{data}
\name{PRSdata}
\alias{PRSdata}
\title{EPIC-Norfolk polygenic risk score data}
\format{
\subsection{\code{PRSdata}}{

A data frame with 50 rows and 5 columns:
\describe{
\item{id}{Anonymous identifier}
\item{age}{Age at recruitment}
\item{sex}{Biological sex 1=man 2=woman}
\item{prs}{A polygenic risk score}
\item{poag}{Primary open angle glaucoma status}
...
}
}
}
\source{
\url{https://www.epic-norfolk.org.uk}
}
\usage{
PRSdata
}
\description{
A subset of data from the EPIC-Norfolk study
}
\keyword{datasets}
