% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_plot.R
\name{annotation_plot}
\alias{annotation_plot}
\title{A function to draw a density plot comparing cases and controls by PRS}
\usage{
annotation_plot(
  PRSdata,
  exposure,
  outcome,
  covariates,
  comparison = NA,
  nquantiles
)
}
\arguments{
\item{PRSdata}{A dataframe which includes an exposure and outcome}

\item{exposure}{A character string containing the name of the exposure variable}

\item{outcome}{A character string containing the name of the outcome variable}

\item{TidyCText}{blah}

\item{DelongROC}{blah}

\item{DelongPValue}{blah}

\item{nq}{Number of quantiles}

\item{TidyQText}{blah}
}
\description{
A function to draw a density plot comparing cases and controls by PRS
}
\examples{
annotation_plot(PRSdata, exposure="prs",outcome="disease", covariates="age,sex", nquantiles=10)
}
