% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectPRS.R
\name{inspectPRS}
\alias{inspectPRS}
\title{A patchwork of plots and PRS summary statistics for an overview of Polygenic Risk Scores}
\usage{
inspectPRS(PRSdata, exposure, outcome, covariates, comparison = NA, nquantiles)
}
\arguments{
\item{PRSdata}{A dataframe which includes an exposure and outcome}

\item{exposure}{A character string containing the name of the exposure variable}

\item{outcome}{A character string containing the name of the outcome variable}

\item{covariates}{Character variable containing names of covariates}

\item{comparison}{Character variable containing name a secondary PRS}

\item{nquantiles}{Number of quantiles}
}
\value{
A patchwork ggplot object.
}
\description{
A patchwork of plots and PRS summary statistics for an overview of Polygenic Risk Scores
}
\examples{
\dontrun{
inspectPRS(PRSdata, exposure="prs",outcome="disease",covariates=c("age", "sex"),nquantiles=10)
}
}
