% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_quantile_plot.R
\name{prs_quantile_plot}
\alias{prs_quantile_plot}
\title{A function to draw a quantile plot comparing cases and controls by PRS}
\usage{
prs_quantile_plot(PRSdata, exposure, outcome, covariates, nquantiles)
}
\arguments{
\item{PRSdata}{A dataframe which includes an exposure and outcome}

\item{exposure}{A character string containing the name of the polygenic risk score}

\item{outcome}{A character string containing the name of the outcome variable}

\item{covariates}{A character string containing a comma separated variable list of covariates}

\item{nquantiles}{Number of quantiles}
}
\description{
A function to draw a quantile plot comparing cases and controls by PRS
}
\examples{
prs_quantile_plot(PRSdata,exposure="prs",outcome="disease",covariates="age,sex",nquantiles=10)
}
