% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_quantile_plot.R
\name{prs_quantile_plot}
\alias{prs_quantile_plot}
\title{A function to draw a quantile plot comparing cases and controls by PRS}
\usage{
prs_quantile_plot(prsModel)
}
\arguments{
\item{prsModel}{A \code{prsModel} object, created by \code{\link[=prs_models]{prs_models()}}.}
}
\value{
A ggplot object.
}
\description{
A function to draw a quantile plot comparing cases and controls by PRS
}
\examples{
\dontrun{
prs_models(PRSdata,exposure="prs",outcome="disease",covariates=c("age","sex"),nquantiles=10) |>
  prs_quantile_plot()
}
}
