% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSdata.R
\docType{data}
\name{PRSdata}
\alias{PRSdata}
\title{Synthetic polygenic risk score data}
\format{
\subsection{\code{PRSdata}}{

A data frame with 50 rows and 5 columns:
\describe{
\item{id}{Anonymous identifier}
\item{age}{Age in years}
\item{sex}{Biological sex 1=man 2=woman}
\item{prs}{A polygenic risk score}
\item{disease}{Disease status 1=case 0=control}
...
}
}
}
\usage{
PRSdata
}
\description{
A synthetic dataset adapted from published data
}
\keyword{datasets}
